LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-iframe-location-expected.txt



FAIL location API on another window fires navigate event in the target window only assert_equals: expected (object) null but got (string) ""
